+++ src/shader/monochrome_shader.rs
@@ -14,7 +14,7 @@ impl Shader for MonochromeShader {
         _surface_normal: Vector3<f64>,
         _world: &World,
         _surface_pos: Vector2<f64>,
-        _recursion_depth: u64,
+        _recursion_depth: f64,
     ) -> Vector3<f64> {
         self.color
     }
